task: refactor — move session concerns to app/, fix dependency inversion (Task 192, Step 2)
The shared auth/session store must not import from a feature. Move the
session concerns out of features/home into app/ (git mv, history kept):

- features/home/logic/types.ts        -> app/logic/current-user.types.ts
- features/home/logic/current-user.api.ts   -> app/logic/
- features/home/logic/dashboard-user.logic.ts -> app/logic/
- features/home/schemas/current-user.schema.ts -> app/schemas/
- the two unit tests                  -> app/__tests__/

Update all importers (auth.store + fetch.api/auth.store tests). The home
feature's logic/ and schemas/ dirs are now empty; HomePage already reads
the store, so no feature consumes these directly. Dependency direction is
features/ -> app/ only.

Completes Task 192 (closes the state-layer half of improvements I1).
Folder moved to ProductSpecification/tasks/done/.

Frontend: 53 vitest passed, 0 failed. Lint (oxlint/eslint/prettier/
vue-tsc) clean. IDE inspections clean on changed files.
