+++ ProductSpecification/tasks/done/192-refactoring-pinia-state-store/decisions/pinia-auth-store-decision.md
