+++ ProductSpecification/tasks/done/192-refactoring-pinia-state-store/progress.md
@@ -16,4 +16,4 @@ Type: refactoring
 - [x] refactor (migrate home subtree) → `HomePage` calls `store.loadMe()` + reads `store.dashboardUser` (local `loading` ref kept — presentational, per frontend-rules; ADR `loading` field NOT added to store); `UserMenu` reads `store.dashboardUser` via `storeToRefs`; dropped the 4-level prop chain (`DashboardShell`/`DashboardTopBar` no longer pass user props); component + vitest suite green → commit
 - [x] refactor (logout via store) → `UserMenu` logout → `store.logout()` + `router.push('/login')`, drop `globalThis.location.reload()`; added store `logout()` action (calls logout api → `reset()`) + store unit test (100% coverage); provided a memory router to `home.smoke` to satisfy `useRouter()` → commit
 - [x] refactor (decouple transport + guard skeleton) → `apiFetch` 401 → `authStore.reset()` (dropped `@/router` import/push); `unauthorized-redirect.logic` split into `isUnauthorized(status)` (transport) + `shouldRedirectToLogin(requiresAuth, isAuthenticated)` (guard); added `router.beforeEach` skeleton honoring `meta.requiresAuth` (no route opts in yet, `RouteMeta` augmented). Updated vitest (`fetch.api`, `unauthorized-redirect.logic`); login.api test gets active Pinia (login 401 now hits the reset path). User decisions: (Q1=A) #162 transport auto-redirect removed by design — deleted obsolete e2e `unauthorized-redirect.spec.ts` (401-on-/activate now renders the Link Expired view, already covered by the 422 expired-token e2e) + dropped its dead `givenSessionExpired` stub; (Q2) logout → `router.push('/')` (welcome), corrected `UserMenu` + ADR → commit
-- [ ] refactor (fix dependency inversion: session types/client → `app/`) → the shared layer (`app/stores/auth.store.ts`) must not import from a feature; move the session concerns out of `features/home/logic/` into `app/` — `AuthenticatedUser`/`CurrentUserResult` (types), `fetchCurrentUser` + `current-user.schema` (api client), and the dashboard view-model (`buildDashboardUser`/`DashboardUser`) the store getter depends on. Update all importers (store, `UserMenu`, tests) and their `__tests__` locations; direction becomes `features/` → `app/` only. Run full vitest + lint → /refactor → commit
+- [x] refactor (fix dependency inversion: session types/client → `app/`) → moved (git mv, history preserved) `features/home/logic/types.ts`→`app/logic/current-user.types.ts` (`AuthenticatedUser`/`CurrentUserResult`), `current-user.api.ts`→`app/logic/`, `dashboard-user.logic.ts`→`app/logic/` (`buildDashboardUser`/`DashboardUser`), `home/schemas/current-user.schema.ts`→`app/schemas/`, and the two unit tests→`app/__tests__/`. Updated all importers (store + `fetch.api`/`auth.store` tests); `home/logic` & `home/schemas` now empty (removed). Dependency direction is `features/` → `app/` only; the shared store no longer imports a feature. Full vitest 53 passed + lint clean → commit
