+++ ProductSpecification/tasks/done/192-refactoring-pinia-state-store/spec.md
