+++ frontend/src/app/logic/current-user.types.ts
